+++ htf/test-py/test_utils.py
@@ -327,27 +327,23 @@ class test_mol_properties(unittest.TestCase):
 
 
 class test_trajectory(unittest.TestCase):
-    def test_run_from_trajectory(self):
+    def test_iter_from_trajectory(self):
         try:
             import MDAnalysis as mda
         except ImportError:
             self.skipTest(
-                "MDAnalysis not available; skipping test_run_from_trajectory")
+                "MDAnalysis not available; skipping test_iter_from_trajectory")
         import math
         import os
         test_pdb = os.path.join(os.path.dirname(__file__), 'test_topol.pdb')
         test_traj = os.path.join(os.path.dirname(__file__), 'test_traj.trr')
         universe = mda.Universe(test_pdb, test_traj)
         # load example graph that calculates average energy
-        model_directory = build_examples.run_traj_graph()
-        hoomd.htf.run_from_trajectory(model_directory, universe,
-                                      period=1, r_cut=25.)
-        # get evaluated outnodes
-        variables = hoomd.htf.load_variables(model_directory,
-                                             ['average-energy'])
-        # assert they are calculated and valid?
-        assert not math.isnan(variables['average-energy'])
-        assert not variables['average-energy'] == 0.0
+        model = build_examples.LJVirialModel(16)
+        for input, ts in hoomd.htf.iter_from_trajectory(16, universe, period=1, r_cut=25.):
+            result = model(input)
+
+        assert np.sum(result[0]) != 0, 'Forces not be computed correctly'
 
 
 if __name__ == '__main__':